+++ streaming_jupyter_integrations/__init__.py
@@ -1,0 +1,15 @@
+import json
+from pathlib import Path
+from typing import Dict, List
+
+HERE = Path(__file__).parent.resolve()
+
+with (HERE / "labextension" / "package.json").open() as fid:
+    data = json.load(fid)
+
+
+def _jupyter_labextension_paths() -> List[Dict[str, str]]:
+    return [{
+        "src": "labextension",
+        "dest": data["name"]
+    }]